ChatEmbed class structure implemented

+++ main.py
@@ -67,6 +67,10 @@ async def on_message(message):
         await mbox.logout()
     
     for profile in profiles:
+        if message.content == 'test':
+            logging.info('Received test from {0.name}'.format(message.author))
+            await profile.messenger.send_gui()
+            break
         if profile.command_channel == message.channel:
             await message.delete()
             await tasks.parser.message(message, profile)
